Parameters

GamsStructure.GamsParameter — Type

GamsParameter{N}(GU,sets::Tuple{Vararg{Symbol}},value::Array{Float64,N},description::String)

Container to hold parameters.

Parameters can be indexed either by set name

P[:set1,:set2]

or by list of element names

P[[:element1,:element2],[:e1,:e2]]

or a mix of both

P[:set1,[:e1]]

GamsStructure.GamsParameter — Method

GamsParameter(base_path::String,parm_name::Symbol,sets,GU::GamsUniverse;description = "",columns = missing)

Load a GamsParameter from a file.

GamsStructure.GamsParameter — Method

GamsParameter(base_path::String,parm_name::Symbol,sets,GU::GamsUniverse;description = "",columns = missing)

Load a GamsParameter from a file. The columns vector is the specific columns to extract from the document. This is primarily useful when the column names of a file either don't match the sets or repeat.

The value column is assumed to be at the end and should not be included in the columns argument.

GamsStructure.@GamsParameters — Macro

@GamsParameters(GU,block)

Create many empty parameters

@GamsParameters(GU,begin
    :P, (:set_1,:set_2), "Description 1"
    :X, (:set_1,), "Description 2"
end)

GamsStructure.@GamsParameters — Macro

@GamsParameters(GU,base_path,block)

Load parameters from a file. This will search base_path\name.csv.

@GamsParameters(GU,begin
    :P, (:set_1,:set_2), "Description 1"
    :X, (:set_1,), "Description 2"
end)

GamsStructure.domain — Method

domain(P::GamsParameter)

Return the domain of the paramter P in the form of a vector of symbols.