Parameters

GamsStructure.GamsParameter — Type

GamsParameter{N}(GU,sets::Tuple{Vararg{Symbol}},value::Array{Float64,N},description::String)

Container to hold parameters.

Parameters can be indexed either by set name

P[:set1,:set2]

or by list of element names

P[[:element1,:element2],[:e1,:e2]]

or a mix of both

P[:set1,[:e1]]

GamsStructure.load_parameter — Method

load_parameter(GU::GamsUniverse,
               path_to_parameter::String,
               domain::Tuple{Vararg{Symbol}};
               description::String = "",
               columns::Union{Vector{Int},Missing} = missing,
               value_name = :value
               )

Load and return a parameter from a CSV file.

GU - Parent universe. The parameter will not be added to the universe

path_to_parameter - Where the parameter lives

domain - A tuple of the names of the domain sets

description - A description of the parameter. Defaults to empty string

columns - If the names in the CSV don't match the set names, use this to specify which columns correspond to the sets.

value_name - The name of the column where the values live. Can also be an integer.

This function requires a specific format of CSV file:

set_1	set_2	value

...	...	...

By default, the function expects the columns names in the CSV to match the set names, however, this can be modified using the columns parameter.

GamsStructure.load_parameter! — Method

load_parameter!(GU::GamsUniverse,
                path_to_parameter::String,
                name::Symbol,
                domain::Tuple{Vararg{Symbol}};
                description::String = "",
                columns::Union{Vector{Int},Missing} = missing,
                value_name = :value
                )

Identical to load_parameter except it includes the parameter in GU.

GamsStructure.@load_parameters! — Macro

@load_parameters!(GU,base_path,block)

Load parameters from a file. This will search base_path\name.csv.

@load_parameters!(GU,begin
    :P, (:set_1,:set_2), description => "Description 1"
    :X, (:set_1,), kwpairs*
end)

kwpairs supported:

description

columns

value_name

file_name - Changes the path to basepath\filename

GamsStructure.@create_parameters — Macro

@create_parameters(GU,block)

Create many empty parameters

@create_parameters(GU,begin
    :P, (:set_1,:set_2), "Description 1"
    :X, (:set_1,), "Description 2"
end)

GamsStructure.domain — Method

domain(P::GamsParameter)

Return the domain of the paramter P in the form of a vector of symbols.